+++ test/add-missing-schema-fields-project/app.js
@@ -1,0 +1,4 @@
+import config from './config.js';
+import apostrophe from '../../index.js';
+
+export default apostrophe(config);